Refactoring quotes for future
