+++ src/handlers/kanyeQuoteHandler.js
@@ -1,8 +1,8 @@
-const {getKanye} = require('../repositories/quoteRepository');
+const {get} = require('../repositories/kanyeQuoteRepository');
 
 module.exports = async () => {
   try {
-    const data = await getKanye();
+    const data = await get();
     const body = JSON.parse(data.body);
     const quote = body.quote;
     return {
